+++ spec/unit/target/pod_target_spec.rb
@@ -349,6 +349,23 @@ module Pod
           exception = lambda { @test_pod_target.product_type_for_test_type(:weird_test_type) }.should.raise Informative
           exception.message.should.include 'Unknown test type `weird_test_type`.'
         end
+
+        it 'returns the correct test type for product type' do
+          @test_pod_target.test_type_for_product_type(:unit_test_bundle).should == :unit
+        end
+
+        it 'raises for unknown product type' do
+          exception = lambda { @test_pod_target.test_type_for_product_type(:weird_product_type) }.should.raise Informative
+          exception.message.should.include 'Unknown product type `weird_product_type`'
+        end
+
+        it 'returns correct copy resources script path for test unit test type' do
+          @test_pod_target.copy_resources_script_path_for_test_type(:unit).to_s.should.include 'Pods/Target Support Files/CoconutLib/CoconutLib-Unit-Tests-resources.sh'
+        end
+
+        it 'returns correct embed frameworks script path for test unit test type' do
+          @test_pod_target.embed_frameworks_script_path_for_test_type(:unit).to_s.should.include 'Pods/Target Support Files/CoconutLib/CoconutLib-Unit-Tests-frameworks.sh'
+        end
       end
     end
   end